requirements.txt
pygame>=2.5.0
python-chess>=1.999
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
pandas>=2.0.0
